Split utility methods into separate file
