+++ pkg/implementer.go
@@ -10,34 +10,6 @@ import (
 	"io/ioutil"
 )
 
-// Get some ordinal ast.Ident.Name from a given ast.Node. A negative will return
-// the last identifier in the tree.
-func getIdent(node ast.Node, ord int) string {
-	ident := ""
-	n := 0
-
-	ast.Inspect(node, func(child ast.Node) bool {
-		if child, ok := child.(*ast.Ident); ok {
-			ident = child.Name
-			n++
-		}
-		return ord < 0 || n == ord
-	})
-
-	return ident
-}
-
-// As a shortcut we parse the receiver expression, then just take the last
-// identifier specified in the resulting ast
-func getType(recv string) (string, error) {
-	a, err := parser.ParseExpr(recv)
-	if err != nil {
-		return "", err
-	}
-
-	return getIdent(a, -1), nil
-}
-
 // An Implementer can, for a certain directory, create and/or update
 // implementation with Go source code for a particular interface
 type Implementer struct {
@@ -54,26 +26,6 @@ type Implementer struct {
 	file map[string]*ast.Package
 	fset *token.FileSet
 	buf  *bytes.Buffer
-}
-
-func getMethods(id string, f *ast.File) []*ast.FuncDecl {
-	decls := []*ast.FuncDecl{}
-
-	for _, decl := range f.Decls {
-		switch decl := decl.(type) {
-		case *ast.FuncDecl:
-			if decl.Recv != nil && decl.Name != nil {
-				for _, r := range decl.Recv.List {
-					name := getIdent(r.Type, -1)
-					if name == id {
-						decls = append(decls, decl)
-					}
-				}
-			}
-		}
-	}
-
-	return decls
 }
 
 // Position returns, if found, the token.Position of the end of the type
